+++ testing/e2e/skr/provisioningservice/internal/model.go
@@ -1,4 +1,4 @@
-package provisioningservice
+package internal
 
 type AccessToken struct {
 	Token string `json:"access_token"`
